Move ArchUnit specs into archunit sub-package
Group the architecture-rule specs under
org.elasticsearch.gradle.internal.archunit so the suite is discoverable
and runnable as a group, separate from the functional tests. Pure move
(git mv + package declaration); no logic changes.
