+++ build-tools-internal/src/integTest/groovy/org/elasticsearch/gradle/internal/archunit/ConfigurationCacheArchUnitSpec.groovy
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-package org.elasticsearch.gradle.internal
+package org.elasticsearch.gradle.internal.archunit
 
 import com.tngtech.archunit.base.DescribedPredicate
 import com.tngtech.archunit.core.domain.JavaClass
